Fix JSON deserialization
