+++ Cargo.toml
@@ -16,7 +16,7 @@ members = [
   "crates/sync/alchemy",
   "crates/crypto",
   "crates/broadcast",
-  "crates/prices"
+  "crates/exchange-rates"
 ]
 default-members = ["bin/iron"]
 
@@ -46,7 +46,7 @@ iron-sync-anvil = { path = "crates/sync/anvil" }
 iron-sync-alchemy = { path = "crates/sync/alchemy" }
 iron-broadcast = { path = "crates/broadcast" }
 iron-tracing = { path = "crates/tracing" }
-iron-prices = { path = "crates/prices" }
+iron-exchange-rates = { path = "crates/exchange-rates" }
 tokio = { version = "1.26.0", features = ["full", "sync"] }
 thiserror = "1.0.40"
 ethers = { version = "2.0.0", features = ["ws"] }
