+++ crates/exchange-rates/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "iron-prices"
+name = "iron-exchange-rates"
 version.workspace = true
 edition.workspace = true
 license-file.workspace = true
@@ -11,10 +11,9 @@ authors.workspace = true
 [dependencies]
 iron-types = { workspace = true }
 
-tauri = { workspace = true }
-thiserror = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
+ethers = { workspace = true }
 tokio = { workspace = true }
 once_cell = { workspace = true }
 async-trait = { workspace = true }
