+++ crates/exchange-rates/src/chainlink.rs
@@ -2,5 +2,5 @@ use serde::Deserialize;
 
 #[derive(Deserialize, Debug, Clone)]
 pub struct ChainlinkFeed {
-    pub address: String,
+    address: String,
 }
