+++ crates/exchange-rates/src/lib.rs
@@ -1,0 +1,15 @@
+mod chainlink;
+mod feed;
+mod init;
+
+pub use init::init;
+
+use feed::Feed;
+use serde::Deserialize;
+use std::collections::HashMap;
+
+#[derive(Debug, Deserialize)]
+pub struct Feeds {
+    // Tuple structure: (network, currency, token symbol)
+    feeds: HashMap<(String, String, String), Vec<Feed>>,
+}
