--- crates/prices/src/error.rs
@@ -1,7 +1,0 @@
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum Error {
-    #[error("feed provider {0} not supported")]
-    UnsupportedFeedProvider(String),
-}
